Got the original files running on linux
with a small change to the api file and including the .so driver file for linux, the amcam camera officially works. I just decided to go with the linux version unless it is impossible for me to do so. Maybe I will update TDC001-Docker for non-linux DLLs with that backend container in the future.