Renamed some directories, Frame Rate Issue
So, it seems like the camera, as it sits, runs a max of 2.9 fps basically. As to why, no idea.

I could optimize this code and use the API that amscope included, but that's not in my priorities. I want this in a container, optimization can happen later.

So, the next few commits will probably be me learning how to send a  uint8 array through my API and just getting these scripts in a container.

I like app.py/qt.py and the image quality they give, so I will take whatever output that it gives and just send it over internet to a remote frontend or something.