backend construction 1
So, I got the amscope camera files moved into the Controllers/backend directory.

I titled the old "tdc_server.py" script to actually reflect the device, which it has been renamed to "amscope_server.py"
(automate this)

I moved the old scripts out of the way for the TDC stepper, and replaced them with the amcam stuff.

I also need to refresh the wheelhouse/ readme.lock directories, and entering a .venv in a directory outside of /Code is annoying, so I would like to make a script for this. (automate)