backend construction
The TDC001-project directory in /Code cannot be renamed without breaking literally all the helper scripts in Linux, MacOS,  Windows.  (fix this on the main directory, rename to "Project" for more general name)

I updated the dockerfile for the backend construction to be a bit more logical (its not perfect, but it works) now, that should be put actually in the original repo since it is a decent upgrade.  (move to original repo)

The wheelhouse and requirements.lock inside the controllers+backend directory for building have to be updated always, so it is probably best to have a script to do the whole update process with wheelhouse instead in one click (future upgrade)