fastapi operational (not neat yet)
just decided to scrap most of what the original developer did and make a fastapi file for this in the end.

The new script can change resolution, set exposure and gain, grab a single frame, return the status, connect/disconnect, all the general functions of a camera.

I made it with chatgpt and debugged it for a bit, but I am disappointed with code readability, so I am going to do future work to get it a bit neater (split up to separate files, add comments, change naming conventions, remove unnecessary syntax)

Anyway, for now, this does all the functions of the camera seemingly without too many errors (edge cases seem fine).