server updates - bugfixes
corrected boot problems with the container, setup.py and launch.py.

The major thing is that this backend container serves an internal port 8000 for the uvicorn server in the entrypoint.sh file, this is also mapped in the docker config to whatever external port you set. Don't mess with it if you don't want things to break lol.

I also changed to get and set notation for the URLs served by the API, as this is more intuitive.

I will try to now do hardware identification in setup.py and figure out how to get it to exclusively map to a device instead.