OS - work in progress
trying to get the setup.py script to be kinda ideal for setting up docker containers, not just for this project, but for multiple. I want to make sure the image builds and only binds to a unique camera each time upon boot without fail.

I don't know the best way to do it that isn't too heavy handed, but ensures the docker image only can send commands to a specific connected device on the system.