All legacy shell scripts moved, fixing images by device serial number in the works
While this code is still mostly broken (due to how it checks hardware ids), the /get_ping function includes a fix which correctly reads a device ID, performs string methods to put it in a correct format, and compares it to the assigned device id in  the device_config.json made by the setup.py script in the "OS" directory (the new directory for interacting easily with this code).

However, the OS directory contains the amcam.py and .so driver due to dependencies being used for this locating (device ids).

This was coded quickly, but the approach is starting to become more apparent as to how to do this, so I will write agnostic scripts which utilize standard usb libraries served by python to do a similar task.

These revisions will be included in the other repository which acts as a base for this project, TDC001-docker (will probably be renamed to "Lab Docker" or something.