windows compatible uvicorn server - docker image not detecting usb devices - not tested on linux
the issue of no usb device detection is still present on windows. While get_ping properly displays devices in the docker container, the remainder does not work. For uvicorn, everything works to my knowledge (running server on host machine). As to why, I cannot figure it out. For today, I am done, so, this will be my last commit for today.

This commit may have broken linux functionality, it is untested and needs to be tested again on linux for compatibility purposes.