V1.0  release complete, ready for first merge
Fixed up basically everything, going to close this project off fast and move to the wave meter for now.

The MU300 was previously working, but an unknown recent issue has prevented it from doing so.

It is unrelated to my repo, but rather wsl and usbipd in windows and binding device to docker container causing the problem. The other camera attached to the microscope (MU503B) works without error.

I will release a patch, but it is not critical for now, as it can be (technically) run in host mode by starting the amscope server locally instead. I may add this functionality to the launch.py script in the future for issues like this.